// src/fsd/entities/Montage/model/types.ts
export type MontageType = {
  title: "Монтаж";
  m_16_20: number;
  m_22_24: number;
  m_26_36: number;
  distance: number;
  m_26_36_hourly: number;
  climber: number;
  complex_fringe: number;
  complex_neon: number;
  complex_thread: number;
  priceObj: {
    m_16_20: number;
    m_22_24: number;
    m_26_36: number;
    distance: number;
    m_26_36_hourly: number;
    complex_fringe: number;
    complex_neon: number;
    complex_thread: number;
  };
};